refactor(listener): extract file command session
